+++ CHANGELOG.md
@@ -28,8 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - MySQL connector
 - Documented style for imports.
 - Documented use of numpy docstrings.
+- Helper utilities to take callables with numpy style docstrings and
+  create config classes out of them using `make_config`.
 ### Changed
 - All instances of `src_url` changed to `key`.
+- Moved from tensorflow 1 to tensorflow 2.
 - `readonly` parameter in source config is now changed to `readwrite`.
 - `predict` parameter of all model config classes has been changed from `str` to `Feature`.
 - Defining features on the command line no longer requires that defined features
@@ -40,10 +43,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use `FastChildWatcher` when run via the CLI to prevent `BlockingIOError`s.
 - TensorFlow based neural network classifier had the `classification` parameter
   in it's config changed to `predict`.
+- SciKit models use `make_config_numpy`.
 ### Fixed
 - CONTRIBUTING.md has `-e` in the wrong place in the getting setup section.
 - Since moving to auto `args()` and `config()`, BaseConfigurable no longer
   produces odd typenames in conjunction with docs.py.
+- Autoconvert Definitions with spec into their spec
 ### Removed
 - The model predict operation erroneously had a `msg` parameter in it's config.
 - Unused imports identified by deepsource.io